Move timedIterations() to SCryptUtil (thanks GrmpCerber)
